pyproject.toml
[project]
name = "cruxpal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.0",
    "fastapi>=0.128.0",
    "greenlet>=3.3.0",
    "opencv-python>=4.12.0.88",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.45",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]
